upodman: new package
This software can be used as a "remote controller" to podman.
Podman already comes with quite large REST API but for my
purposes I needed to have some changes in resulting structure
and in the way that things are done. Also, some of actions
should be executed in background as there always is not
enough time until receiving process times out, so that's
what this program mostly does.

One potential use case would be container management in LuCi.
LuCi already has a quite comprehensive add-on for docker which I
ported to support podman (all but load and ram stats that caused
some unexpected difficulties) for most parts and added POD support.
Still, I think that full container management is best done in a
terminal with provided tools and I loved how it was done
with Cockpit, so it inspired me and I started to build something
of my own based on it and from that idea, upodman was born.

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>